requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
pymongo==4.6.0
sqlalchemy==2.0.23
pandas==2.1.3
loguru==0.7.2
aiohttp==3.9.1
asyncio==3.4.3
webdriver-manager==4.0.1 